implement async fetch feature
using `taskqueue' provided by GAE to implement async fetch feature.
Whenever the app wants to fetch a url, it would add the url to taskqueue,
then the task in taskqueue will trigger the `fetcher worker'.
The fetcher worker use `urlfetch' to fetch the url and parse its content, then save it to Datastore.